Devil dog mode and meter
press Q to access devil dog mode. will become invulnerable and no ammo consumption. when the meter drains to 0 the effect is over.

+++ Echoes of Belleau/Assets/Scripts/gameManager.cs
@@ -396,7 +396,7 @@ public class gameManager : MonoBehaviour
         currentWeaponIcon.enabled = (icon != null);
         currentWeaponIcon.sprite = icon;
     }
-    public void updateDevilDogBar(int currentPoints, int maxPoints)
+    public void updateDevilDogBar(float currentPoints, float maxPoints)
     {
         if(devilDogBarFill != null)
         {
